requirements.txt
Flask==2.0.1
flask-cors==3.0.10
supabase==1.0.3
PyJWT==2.3.0
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.3 